pyproject.toml
[tool.poetry]
name = "emotion_analysis"
version = "0.1.0"
description = ""
authors = ["cbr9 <cabero96@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.2"
hydra-core = "^1.3.0"
transformers = "^4.25.1"
torch = "^1.13.1"
pytorch-lightning = "^1.8.5"
hydra-zen = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
